requirements.txt
py-trello
nextcloud-deck
pyyaml
nextcloud-api-wrapper
